*** empty log message ***
git-svn-id: svn://svn.code.sf.net/p/actiongame/code/trunk/ac@687 e016b1c5-ed68-44aa-a2f7-31cac88cffde

+++ source/src/clientgame.cpp
@@ -299,7 +299,7 @@ int fixspawn = 2;
 // returns -2 for a free place, else dist to the nearest enemy
 float nearestenemy(vec place, char *team)
 {
-    int nearestenemydist = -1;
+    float nearestenemydist = -1;
     loopv(players)
     {
         playerent *other = players[i];
